train and test MLP on tank system, but no good results

+++ learning-qp-master/rl_games/rl_games/algos_torch/network_builder.py
@@ -263,7 +263,7 @@ class A2CBuilder(NetworkBuilder):
                 sigma_init = self.init_factory.create(**self.space_config['sigma_init'])
 
                 if self.fixed_sigma:
-                    self.sigma = nn.Parameter(torch.zeros(actions_num, requires_grad=True, dtype=torch.float32), requires_grad=True)
+                    self.sigma = nn.Parameter(torch.zeros(actions_num, requires_grad=True), requires_grad=True)
                 else:
                     self.sigma = torch.nn.Linear(out_size, actions_num)
 
@@ -647,7 +647,7 @@ class A2CResnetBuilder(NetworkBuilder):
                 sigma_init = self.init_factory.create(**self.space_config['sigma_init'])
 
                 if self.fixed_sigma:
-                    self.sigma = nn.Parameter(torch.zeros(actions_num, requires_grad=True, dtype=torch.float32), requires_grad=True)
+                    self.sigma = nn.Parameter(torch.zeros(actions_num, requires_grad=True), requires_grad=True)
                 else:
                     self.sigma = torch.nn.Linear(out_size, actions_num)
